requirements.txt
# Dependencies for this project
chartbook[all]>=0.0.3
doit==0.36.0
pandas==2.2.3
numpy==1.26.4
pyarrow
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.12.0
jupyter
jupyterlab
notebook
ipynb-py-convert
myst-nb
sphinx==7.3.7
sphinx-book-theme==1.1.3
python-dotenv==1.0.1
python-decouple
requests
wrds==3.2.0
